Moved initialization of system dataset modules to DatasetTypeService, so they can be governed by authorization enforcement rules as well, just like other entities.
